:lipstick: Update science badge URL and color
The other badges in the same row have Wikipedia links. This commit will
update the science badge to also point to Wikipedia. This commit will
also change the color from red ("critical") to green ("success"). The
red color was intended to represent that science was critical, but it
could also be interpreted to mean failure (as in a failed CI/CD run).